Updating NPC to be compatible with Cacti 1.x
This is the first update of many I think.  Right now, we just want to
get this working with Cacti 1.x.  There are a number of additional
changes, and this may not work 100%.

It looks like Billy was working on making the ExtJS panels sortable when
the project stalled.  I've tried to carry those changes forward, but may
have missed a few.  So, let's us know what is broken.

I've removed the incompatible ExtJS framework adapters to allow
licensing to stay in tact for the near term.